Collapse directory structure
Rails3 requires that the directory that holds the application has a
meaningful name which cannot be "application" as we have currently.
Move application/ contents into the root to simplify things and to
meet this requirement.

The config and spec files were changed accordingly. Unfortunately
upgrades will need to be done with care as the path to the
configuration file has changed.